Radio button filtering working. Bootstrap CSS hides them :(
So not using bootstrap.min.css on the product page yet. Tried many
things to override the formatting, without success
The filtering is done on the server side. I added a method in product
service getProductsByCat. This only downloads the products from the
database if the local stored product list is still empty