Login and out working. User stored in session
I had to add a redirect to the success page in the popup which has a
button to close it. When the popup closes the page you came from is
refrehed and shows the user and logout option in the nav bar.
If the login is unsuccesful the popup windows asks user to try again