Merge Rutger en Jurjen, adding, deleting to cart working, updated format
of products page. Cart controller seriously updated. No dummy user will
be stored, not neccasary. if no user cart a cart is made with attached
subOrder ArrayList. de cart is stored in the session. When checking out
if no user is logged in a new Address is made and stored in the session.
if a user is logged in the user adress will be stored under the same
name áddress'in the session.